+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - `br`: added retained taxes CSLL, INSS and IRRF
 - `pt-saft-v1`: added extensions to handle integration of documents (other systems, manually issued or recovered)
 
+### Fixed
+
+- `fr`: handle exception cases in Tax Identity Codes.
+
 ## [v0.220.6] - 2025-08-12
 
 ### Changed